Refactored Pagination buttons to a simpler and more optimized state
